Add showerthought_cache script
Running the 'showerthought' script at every login, along with any other scripts that might be in update-motd.d is too slow for my liking.
The showerthought_cache script will be run by cron or similar, as often as the user wants, and retrieve a specified number of "thoughts", and store them where the 'showerthought' script, whether run manually or as part of a dynamic MOTD, can choose a random one to display.